Fix DescendantOf handling for transitive descendants.
If a type A inherits from type B, and type B inherits from type C,
it's possible that when analyzing type A that type C isn't present in
the compiler's memory.  To remedy this, when checking if a particular
type is a subtype of another named type, we need to recursively walk
all of the types supertypes and interfaces to ensure they're loaded in
memory so we can do a proper subtype check.