Provide suggested fixes even with -Xjcov off.
This change reworks the error message generation so that suggested fixes
("Did you mean...?") are provided even when tree end positions are not
available.  When error-prone wants to print an error message, if tree
end positions are not available, we reparse the file and store the end
position map for later use.

This should have no performance cost in the common case when error-prone
does not detect any errors.