Add MalformedFormatString check.
This check looks for calls to printf-like methods in which the format
string does not match the arguments in the method call.  As a first
step, it only flags the case where there are extra arguments, so that we
can provide a safe fix that simply deletes the extra args.  Eventually
this will detect other format-string-related bugs.