Make error-prone compatible with JDK8.
The 'jdkcompat' project defines an interface that abstracts over
internal APIs that changed in JDK8, and the 'jdk8' and 'jdk7' modules
implement versions of the inferface that work against the different
versions of the JDK. (JDK6 is still supported, but doesn't require
separate handling from JDK7).

When error-prone is built the current JAVA_HOME setting is used to
determine which jdkcompat library to use.