==Fixing issue 222
We had two issues here:
1- Class cast exception because sometimes the method we are looking at is not a
field access.
2- Index out of bound exception because sometimes type parameters are missing
for a type.

I fixed this issue by introducing a new function which extracts type parameters
from a type and not an expression. Why? Because when we are trying to get the
type parameters of the instance object of a method that is not field access, we
have no expression to start with. However, we do have type information about its
owner. This change will take care of class cast exception.

For index out of bound exception I had to change the code for the new
genericTypeOfType and old genericTypeOf to check if the number of type arguments
is zero. In that case we return the java.lang.Object type from symbol table as
the type argument's type.