Add experimental check for varargs/array mistakes due to boxing.
This check flags any place where a primitive array is passed as a varargs
parameter and is autoboxed into an array with a single element of
primitive array type.  This grew out of our ArrayHashCode check; we
wanted to know if it is *ever* a good idea to do this.  It needs more
tuning before having the maturity bumped to MATURE.