Synchronize the set of checks run in open-source error-prone with the set from internal error-prone.
Sets maturity = MATURE for all checks that are on inside Google and
maturity = EXPERIMENTAL for checks that are not.  Sets severity = ERROR
if it is on as a compiler error and severity = WARNING it is only on at
code review time.

-------------
Created by MOE: http://code.google.com/p/moe-java
MOE_MIGRATED_REVID=76736418