In JavacErrorDescriptionListener, filter nulls from the list of applied fixes, after being transformed from unapplied fixes.  The applied fix can be null if it only affects imports.
For example, consider a checker that just switches out one API for a
compatible one.  If there is a static import for a method from the old API,
then error-prone will match on the methodInvocation, but the fix will only
update the static import and not modify the method call itself.  In this case,
we want the following behavior:
1) In the diagnostic message, we do not want to provide a fix snippet, because
the fix is not local to where the caret will be (on the method invocation).
2) In automated fixes, we do want to update the static import but not
change the method call itself.
-------------
Created by MOE: http://code.google.com/p/moe-java
MOE_MIGRATED_REVID=81279147