Standardize list of on-by-default checks
Currently we have different methods for enabling checks internally and
externally, and this leads to skew for which checks are on in each
environment.

Note that this doesn't fix the problem for warnings checks, only errors.
Those still need to be synced manually.

RELNOTES: Synchronize list of on-by-default checks for open-source Error Prone
and internal Error Prone.
-------------
Created by MOE: http://code.google.com/p/moe-java
MOE_MIGRATED_REVID=90828929