Add logic to not use integral suggestions when using BigDecimal.valueOf().
This is because BigDecimal.valueOf(1.0) => new BigDecimal("1.0") which
has a different scale than BigDecimal.ONE. This is not usually an issue
but it affects equality of BigDecimals which means that it causes tests
that rely on the old scale to fail.

RELNOTES:
Changes BigDecimalLiteralDouble to avoid suggestions with minor semantic
differences.
-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=122649770