New check to remove unused imports
Add a check that removes imports for symbols that are not being used.
In doing so, fix some issues with StaticImports so that it handles the
case where a single static import imports multiple symbols.

RELNOTES: Add a check that removes imports for symbols that are not being used.

MOE_MIGRATED_REVID=124008239