Add a test for ProtoStringFieldReferenceEquality
and then fix the check so the test passes, since it was broken by
b1d78673f73af9ee158534ff5c406997993f8e98.  AbstractReferenceEquality
expects the predicate to match both sides of the comparison, but in this
case we want to match if either side is a proto string field.

MOE_MIGRATED_REVID=127222107