Don't generate hasFoo() suggestion that won't compile.
If you compare the result of a getFoo() call on a proto to null,
ProtoFieldNullComparison suggests calling hasFoo() instead.  But in proto3,
hasFoo() methods are not generated for scalar types.  The warning is still
correct (the result of getFoo() cannot be null), but the suggestion won't
compile.

This CL changes the suggested fix logic to first check whether a hasFoo()
method exists before suggesting it.  If the method does not exist, it does
not suggest a fix.

MOE_MIGRATED_REVID=128416640