Update OverlappingQualifierAndScopeAnnotation
to compile-error on the class that has both annotations (not suggesting
the removal of any particular annotation), as well as provide some more
explanatory flavor on why this pattern is a bad idea.

RELNOTES: Improve documentation on
OverlappingQualifierAndScopeAnnotation, and no longer suggest the
removal of Qualifiers or Scope annotations from the annotation (allowing
the developer to make the choice for themselves).

MOE_MIGRATED_REVID=131022884