Update the interface contract of MultiMatcher
to return all the matching nodes, as opposed to just one of the matching
nodes, and use this new contract in MoreThanOneScopeAnnotationOnClass.

Also, adjusts the behavior of MoreThanOneScopeAnnotationOnClass to
ignore Dagger components/subcomponents, and no longer suggest the
deletion of any particular scope annotation.

RELNOTES: Adjusts the behavior of MoreThanOneScopeAnnotationOnClass to
ignore Dagger components/subcomponents, and no longer suggest the
deletion of any particular scope annotation.

MOE_MIGRATED_REVID=131232952