Delete isCastableTo matcher
It crashes on nulls, and is not used correctly by any checks.

As described in JLS 5.5.1 a class type S is castable to T if S < T *or* T < S.
For example, an instance of static type Object is castable to an instance of
static type String, because the runtime type of the first might also be String.
Checks using this matcher generally wanted to test assignability instead.

MOE_MIGRATED_REVID=131499298