Refactor a bit of ScopeOrQualifierAnnotationRetention
specifically ignoring it for Android builds, as they are likely to use
compile-time DI.

RELNOTES: Restrict ScopeOrQualifierAnnotationRetention to non-Android
builds, as dependency injection frameworks used on Android (e.g. Dagger)
tend to be compile-time, and are thus practically unaffected by the
change. It is still incorrect to have @Scope/@Qualifier withou runtime
retention by spec, but this prevents warnings in circumstances we know
to be OK.

MOE_MIGRATED_REVID=133310049