Make ImportStatements line break handling more portable
This logic failed to trim the trailing `;` from imports with CRLF line
endings, so we were emitting extra semi-colons. When javac see:

import a.A;;
import b.B;

It accepts the extra import as a top-level type declaration due to a
parser bug [1], and then adds all subsequent imports to the list of type
declarations. Finally, compile-testing was searching the list of actual
type declarations and asserting it contained exactly the expected types,
and failing because of the unexpected import trees.

[1] http://mail.openjdk.java.net/pipermail/compiler-dev/2013-August/006968.html

MOE_MIGRATED_REVID=136861403