Add findUnusedIdentifiers and findReferencedIdentifiers methods to FindIdentifiers
findUnusedIdentifiers walks up the tree from the current position
looking for uses and declarations of variables. It returns all variables
which are declared and not used by the current position.
findReferencedIdentifiers walks down a tree collecting a set of symbols
which are referred to.

RELNOTES: Add findUnusedIdentifiers and findReferencedIdentifiers methods to FindIdentifiers

MOE_MIGRATED_REVID=146015029