Update UnnecessaryDefaultInEnumSwitch to cover more cases
if the switch statement is exhaustive and the switch *can* complete
normally, then try merge the default into the prior case or delete it
entirely.

RELNOTES: Update UnnecessaryDefaultInEnumSwitch to cover more cases: if
the switch statement is exhaustive and the switch *can* complete
normally, then try merge the default into the prior case or delete it
entirely.

MOE_MIGRATED_REVID=153858536