Change QualifierOnMethodWithoutProvides
to be a much-lower scope checker: QualifierOrScopeOnInjectedMethod.

Many instances of qualifiers on non-@Provides methods are when normal
annotations (that don't need to be @Qualifiers) are used reflectively in
some non-injection scenario, or are used in some weird inheritance
hierarchy for Dagger modules.

Instead, attempt to only check for instances where @Inject and some
@Qualifier or @ScopeAnnotation annotation are used together on a method:
here, the intent was clear (applying scope/qualifier to the type), but
the mechanics were wrong.

RELNOTES: Replace the disabled QualifierOnMethodWithoutProvides check
with QualifierOrScopeOnInjectedMethod.

MOE_MIGRATED_REVID=154468557