Checker for parameter order on autovalue constructor call
The check works by looking at the arguments used to call the constructor of the AutoValue object. If a different permutation of the arguments would exactly match the parameters of the constructor then a warning is generated. Arguments for which a name cannot be extracted are left in their original position. Occasionally an author will deliberately swap the order of the constructor call e.g.

 if (first < last) {
   return new AutoValue_Foo(first, last);
 }
 else {
   return new AutoValue_Foo(last, first);
 }

We avoid matching on these by looking to see if the permutation of the arguments exists in another call within the same method.

The check is implemented as a special case of the ArgumentSelectionDefectChecker which in general looks to see if a permutation of the argument names is more suitable than the current one.

RELNOTES: Checker for parameter order on autovalue constructor call

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=155635122