Optimize string concatenations within `StringBuilder.append()` calls
Concatenating strings as an argument to `StringBuilder.append()` causes
a new, temporary StringBuilder to be created just for that argument on
modern JVMs, which is a little wasteful. This commit replaces those
concatenations with separate calls to `StringBuilder.append()` so as to
make sure that only one StringBuilder is created, rather than
(1 + concat-args) StringBuilders.