Create WakelockReleasedDangerously checker.
If Android WakeLock `release` is called after the OS has already
released the wakelock  due to timeout, a RuntimeException is thrown.
Check that wakelock.release is always called inside a
try/catch(RuntimeException) block to prevent crashes, *if* that wakelock
is known to be acquired with a timeout, and it is reference counted.

RELNOTES: Add Android WakelockReleasedDangerously checker.

MOE_MIGRATED_REVID=166403692