Allow long non-file arguments to be spilled
Allow long non-file arguments (e.g. -classpath) to go to a file.  This
prevents E2BIG when the classpath argument is too long.

This can't easily be fixed in upstream ant because:
(a) it would require an API change, and
(b) there is no way for ant to know which arguments are safe to spill

Fixes #671

RELNOTES: N/A

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=167726515