Handle ExpectedException tests that end with an explicit failure
This code may be written to ensure the test fails if the code under test
doesn't throw. It's unnecessary and will break the test if the code under
test is wrapped in assertThrows/expectThrows, so remove it.

RELNOTES: N/A

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=174253158