Support more kinds of fail() in CheckReturnValue
This adds support for JUnit 5's fail(), testng's fail(), and throw new
AssertionError().

Using assertThrows() is preferable to using fail() in JUnit 5 and testng,
and assertThrows() is already supported by CheckReturnValue, so mention
that in the docs.

This commit intentionally does not add tests for the new fail() methods
to CheckReturnValueTest.  That would require adding JUnit 5 and testng as
Maven dependencies, which would be confusing because the project already
depends on JUnit 4.

Fixes #804