Improve immutable diagnostics
Diagnostics on non-final or mutable fields are reported on the tree node
for the field if one is available. Unify the logic for those diagnostics
with the regular case, and ensure that there is always context about why
we're checking the field for immutability. This improves diagnostics for
classes that are checked because of a super-type is annotated.

MOE_MIGRATED_REVID=176436334