Remove the original ImmutableSet-using method signatures in ThreadSafety.
This is part 2 of 3 of the binary compatibility cleanup in this class. The last change will be to remove threadSafeTypeParameterSetInScope once previous builds of Error Prone are no longer an issue, and we have safely updated the internal callers.

RELNOTES: N/A

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=180816001