{Equals,Predicate,Collection}IncompatibleType
now explore type parameters to ensure their compatibility: Foo<A> is
incompatible with Foo<B> if A and B are incompatible with each-other.

RELNOTES:
EqualsIncompatibleType/PredicateIncompatibleType/CollectionIncompatible
now explore type parameters to ensure their compatibility: Foo<A> is
incompatible with Foo<B> if A and B are incompatible with each-other.

MOE_MIGRATED_REVID=181215563