Add newlines between diagnostics when printing "All errors", so that it is easier to read them in test log
The "Did you mean" messages are on the same line as the start of the next
message, making them hard to separate.

This change prints the diagnostics without the `[,]` from the list, and puts a
blank line between the diagnostics.

RELNOTES: Make the failure message in DiagnosticTestHelper easier to read.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=182543853