Update maven-javadoc-plugin to 3.0.0.
This is necessary to build with JDK9.
Or rather, it *would* be necessary if not for your override of the plexus-io dependency. But the override feels like something that would break eventually, so it seems worth simplifying this while I'm updating our other projects to 3.0.0.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=182817606