Allow @Immutable to be applied to type parameters on classes
If it is, those type parameters are allowed anywhere that a type parameter
in the containerOf spec is allowed in the declaration of an immutable type.

Additionally, instantiations of classes with @Immutable type
parameters are checked to ensure the type argument is immutable.

RELNOTES: N/A

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=183910034