Improve LazyInit's documentation
There are several issues with the current documentation that I am attempting to address with this CL:

1. before this CL, hashCode is used to exemplify the situations where it's ok for an initialization to be racy; but hashCode's use of a local variable is evidently needed because it is not atomically constructed in this method, which masks the fact that, even if the hashCode was calculated atomically, a local variable would also be required to avoid the problem of out-of-order reading of the (non-volatile) field; this is aggravated by the fact that the javadoc does not state the problem

2. also, using hashCode as an example precludes demonstrating how to apply the @LazyInit annotation (which is not present in String.hash).

To address these problems:

a. the example code has been change to be minimal and self-contained, and provide a template for people to copy-and-paste when using the annotation; this example code now also includes the field that is annotated with @LazyInit

b. a caution has been added to say that the (weird) pattern is necessary, and a link has been added for the extensive documentation

c. the note on "hashCode" has been left, but it simply links to the code in String.java, rather than copy and paste it

RELNOTES: None

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=194325578