Make the ExpectedException refactoring tool work when we can't determine the expected exception class.
ATM, when the ExpectedException refactoring tool encounters a «thrown.expect(...)» it assumes that it's of the form «thrown.expect(SomeClass.class)» and crashes otherwise.
This is definitely a common pattern, but it's not the case in general.
When refactoring some of our tests that don't follow this pattern, the tool failed on some targets.

RELNOTES: Make the ExpectedException refactoring tool work when we can't determine the expected exception class.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=196167112