Add an ErrorProne check for usage of Truth with likely reversed actual/expected values based on the argument to assertThat containing the word "expected" or being a constant.
A suggestion is not generated if the expected value is a compile-time constant; this might lead to some false negatives, but it dealt with a huge class of false positives.

RELNOTES: [TruthAssertExpected] Check for assertions on the expected rather than actual value.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=197122363