Ensure that we're creating a literal with exactly the right value.
If one writes an assertion like,
  assertThat(myDouble).isWithin(0).of(0.1f);
We were rewriting that to,
  assertThat(myDouble).isEqualTo(0.1);

Which will fail, as 0.1 is not exactly representable as either a float or a double. The correct replacement is (double) 0.1f (or an unpleasantly long decimal expansion).

This didn't turn up before because it's a pretty uncommon pattern; the author needs to be asserting equality on a double which came from a float.

RELNOTES: N/A

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=198848809