Add explicit checks for @SuppressWarnings("BadImport") while walking the tree looking for occurrences to fix.
BadImport requires explicit checking of @SuppressWarnings because it matches on the ImportTree, but fixes the occurrences. Since you cannot apply an annotation to an import (easily), it is just more pleasant to be able to apply it to trees containing the usages.

RELNOTES: Allow suppression of BadImport

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=206784964