Remove slow fallback to JavaCompiler#resolveIdent in VisitorState
this doesn't work for modular compilations, and fixing it would require
expensive iteration over the available modules.

The only real use-case for loading symbols that haven't already been
loaded is checking if a library is on the classpath before emitting
fixes, but the suggested fixes already have a strong bias towards using
guava so this is an acceptable regression.

MOE_MIGRATED_REVID=207741811