Create an ErrorProne that removes empty else branches.
This is useful in the nano2lite migration because empty else branches do sometimes get generated, due to AddNullGuards and ClearRemover.

AddNullGuards can lead to code such as:

if (x != null) {
  p.setX(x);
} else {
  p.clearX();
}

ClearRemover can subsequently see that clearX() is unnecessary and leave:

if (x != null) {
  p.setX(x);
} else {
}

A further pass to remove the empty else would make the code cleaner.

This itself is rather generic, so I placed the code in refaster/cleanups. But I can move it to a nano2lite specific location if there's not much desire for such a checker outside of nano2lite.

RELNOTES: N/A

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=208912804