Add SparseArray types to UndefinedEquals, as they do not implement equals.
SparseArray is a bit different in that it's a concrete class not an interface, but it's non-final so we can't be *sure* (in general) that someone hasn't overridden it to provide equals/hashCode. But they almost surely won't have, so special-casing it seems reasonable.

RELNOTES: Add SparseArray to UndefinedEquals.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=220602121