RELNOTES: Disallow references to methods with @CompileTimeConstant parameters.
We are currently not able to check if the method reference is later applied to a @CompileTimeConstant thus we are banning getting references to methods accepting @CompileTimeConstant.

This is controlled by the CompileTimeConstantChecker:ForbidMethodReferences flag for the time being.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=222947017