Omit empty `()` from marker annotations in MultiVariableDeclaration
The fix relies on pretty-printing the desugared AST nodes for the individual
variable declarations. Normally we try to avoid pretty-printing and manipulate
source text directly, but in this case the alternative is to re-implement
parsing and desugaring multi-variable declarations.

RELNOTES: N/A

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=224253024