Fix FormatMethod and FormatString javadoc.
According to https://reflectoring.io/howto-format-code-snippets-in-javadoc/#code-markup-features-at-a-glance

There is just no way to to correctly render "@" using "<pre>{@code ...}</pre>". The only correct way to do it would be with vanilla "<pre>...</pre>" and escape '@' '<' '>' that needs to be escaped (which in this case, is only '@').

RELNOTES: Fix FormatMethod and FormatString javadoc.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=228208840