Fix calls to TreeScanner.scan(TreePath, param).
This is probably not what was intended: because TreePath implements Iterable<Tree>, this will scan from all nodes in the TreePath.
It don't think this is a problem in either case, because they only look to be called with a path referring to a CompilationUnitTree; but this is a bit brittle.

(An alternative fix would be to make these scanners into TreePathScanners. That TreeScanner and TreePathScanner do very different things with a TreePath parameter is quite insidious.

RELNOTES: None

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=228960415