Fix ModifiedButNotUsed doc
The code snippet was calling ".toBuilder" on something that was already a Builder (which I’m 99% sure does not compile?).

The documentation "or the Builder modified in place" which sounds to me like it implies it is a preferred solution. So should be tagged as ".good".

Finally, renamed "setFoo" to "withFoo", as that seems like the convention for functions that "return a modified copy, but don’t touch input". Added "@CheckReturnValue" to drive home the point that if you are calling this method, you still need to *also* use the returned value.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=231218053