Constrain toType's signature to forbid using it in silly ways
To adapt a Matcher to another's type, it's not enough for them to have Tree as a common
ancestor. It would make no sense to, for example, convert a Matcher<ExpressionTree> to
a Matcher<MethodTree>. The instanceof check would fail for all MethodTree objects,
meaning the ExpressionTree match would never be tested.

Instead, we require that the input Matcher be for some subtype of the desired output
type. This way, it is possible that the instanceof check will succeed, causing the
wrapped Matcher to actually exercise its logic.

This does exclude one possible way someone could use toType: if I have a Matcher<Tree> that matches some subset of ExpressionTree and MethodTree nodes, and I want to constrain it to only match MethodTree nodes, it was previously possible to use toType to narrow the scope of this matcher; my proposed change makes it only possible to widen scope. Currently it seems nobody is using this feature, and I think it's a confusing way to use the method, so I propose getting rid of it, but would welcome contrary feedback. After this CL clients would instead need to use isInstance to do this narrowing.

RELNOTES: none

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=231826073