Ensure that Rx return values are not ignored. If it were to be ignored, it would 99% of the time be a bug.
Initially I had the fix as the default "delete the line", but realistically that isn't the usual fix. It is possible that:

They are using it validly and should add a @CanIgnoreReturnValue if:
- The function does some synchronous before returning an Observable
- The function may return an observable and schedule it itself (via connect() etc)

And then even if it is a bug then the programmer probably would want to keep the line and perform some operation onError/onSuccess.

In no situation is deleting likely the right call.

Here is a CL generated via flume with the delete fix just to show that the bugpattern works properly: []

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=233458873