Prepend a supertype check in CollectionIncompatibleType so we can abort quickly if none of the applicable types match.
For each method invocation, we're currently checking for 20 methods in the worst-case, which means 20 type-hierarchy checks. This is costly given that it runs on every method invocation in the compilation. The 20 methods are on types that share one of 3 collective supertypes: adding a first order check for one of them reduces the type-hierarchy checks considerably for the common case. The non-common case should not be affected greatly.

An alternative would be to index the common AbstractCollectionIncompatibleTypeMatchers by supertype first. That may be a little easier to maintain, but my guess is that it would be a little slower for the common case since we'd be reducing 20 -> 7 instead of 20 -> 3. We could try a layered map, but I'm not sure if that's worth it.

This reduces the cost of CollectionIncompatibleType by 45% on some builds.

RELNOTES: Performance improvements

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=248245403