Make MatchState.paramTypes() a derived property to save a costly allocation in a type that is created very frequently.
The ImmutableList.copyOf() is 70% of the current runtime of this method. Reducing it to use Collections.unmodifiableList() cut that time in half, but added its own cost (albeit smaller). Making it a derived property reduces the cost entirely and shifts it to the caller side.

Because the list is never modified by javac and we're in a single-threaded environment, and because this is an internal type with only one caller, I didn't bother retaining the ImmutableList type from the now-derived property method. But I can add that back if we'd like.

This is used throughout ErrorProne, most notably perhaps in CollectionIncompatibleType, which is the 2nd most expensive checker internally.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=248255462