Add a cache around ASTHelpers.isInherited
This is called quite often, and with a very small number of distinct
arguments. I'd like to speed up the actual implementation, or avoid
calling it so often, but for starters introducing a cache speeds up
the method by around 80%, shaving 2% off of the entire javac action.

This adds a new dependency on Caffeine; I considered using a Guava
Cache since we already depend on Guava, but Caffeine's version is 3
times as fast for this use case.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=259064453