Add support for inline caches
As the javadoc suggests, these can be used to memoize an expensive
operation that requires a VisitorState but is not expected to change
throughout the compilation.

To manage the lifecycle of these caches, also add a new item to the
Context representing compilation context.

Also, as a proof of concept, add one inline cache, in front of
Suppliers.fromString - this alone halves the time spent converting
String to Type in the entire compilation.

More strategically-placed caches to come in future.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=262249676