ArrayToString: don't complain about @FormatMethods. They may be doing something special to handle arrays correctly.
This was changed to start checking @FormatMethods. Unfortunately(?), it's quite possible to annotate a @FormatMethod which is more clever than the average, which leads to false positives. I'm backing away from that change just to get this re-enabled as an error.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=279288347