UnnecessaryCheckNotNull: be more careful about qualifying the precondition we want to use.
And encourage static importing it; that is the Java practices guidance after all.

This introduces a SuggestedFixes helper to get a name to refer to methods that we want to try to static import, but fall back to qualifying if necessary.

This addresses PR #1433, though not in the same way.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=288470257