Fix a bug when determining whether a formal parameter is varargs, and the parameter/method is declared in a separate compilation.
The VARARGS javac flag is intended only to be set for methods, not parameters.
It seems to be set for parameters that we have as source, but not for ones
where we only have the bytecode.  So the correct way to check if a formal
parameter is varargs is to check if the owning method is varargs and the
parmeter is the last one.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=289470666