RestrictedInheritance/RestrictedApi: ignore the checkerName in the annotations.
Excitingly, it's critical that this /does not/ match the #canonicalName defined by the BugPattern. These checks rely on being able to downgrade the check to a WARNING based on a value in the annotation. If the finding were reported under the canonical name, the severity would be overridden back to ERROR.

This is terrible, but I've left a TODO to clean this up.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=300297280