Demote severity of ExtendsAutoValue and RefersToDaggerCodegen to a warning
Fixes: #1619.

Both checks appear to be a breaking changes for Bazel EP integration.

The way Bazel cross compilation works is a bit convoluted, but the
reason for false positive boils down to: the missing annotation
generation is using -source 8 -target 8 with a JDK 11 -bootclasspath
(because it's targeting the remote JDK 11). So AutoValue processor is
trying to use javax.annotation.Generated based on the source level, but
it isn't available because the JDK 11 bootclasspath only contains
javax.annotation.processing.Generated.

Demote the severity to warning for the checks that rely on having JDK
8's: javax.annotation.Generated or JDK 11's
javax.annotation.processing.Generated on the classpath today, and
consider to promote it to error again later, when Java 8 is gone.

That would substantially simplify the adoption of new EP releases,
because it wouldn't require the addition of javax.annotation library
to the build tool chain.