Replace TODOs with previously-disallowed usernames with bug numbers.
With MOE, we had an allowlist of usernames to publish in TODO comments.  It did not include these users, so they were replaced with "TODO(user)".  Copybara has similar functionality, but it is awkward (see b/153568632) and we aren't using it in Java team projects.  For now, I'd like to replace the usernames with bug numbers to avoid publishing them when we switch to Copybara.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=316197659